+++ test/PHPUnit/type-dossier/transformation/TypeDossierTransformationTest.php
@@ -1,6 +1,7 @@
 <?php
 
 require_once __DIR__ . "/../../pastell-core/type-dossier/TypeDossierLoader.class.php";
+require_once __DIR__ . "/../../../../connecteur/glaneur-sftp/GlaneurSFTP.class.php";
 
 class TypeDossierTransformationTest extends PastellTestCase
 {
@@ -10,6 +11,12 @@ class TypeDossierTransformationTest extends PastellTestCase
     /** @var TypeDossierLoader */
     private $typeDossierLoader;
 
+    /** @var TmpFolder */
+    private $tmpFolder;
+
+    /** @var string */
+    private $workspace_path;
+
     /**
      * @throws Exception
      */
@@ -17,12 +24,18 @@ class TypeDossierTransformationTest extends PastellTestCase
     {
         parent::setUp();
         $this->typeDossierLoader = $this->getObjectInstancier()->getInstance(TypeDossierLoader::class);
+        // pour le glaneur:
+        $this->tmpFolder = new TmpFolder();
+        $this->workspace_path = $this->tmpFolder->create();
+        $this->getObjectInstancier()->setInstance('workspacePath', $this->workspace_path);
     }
 
     public function tearDown()
     {
-        parent::tearDown();
         $this->typeDossierLoader->unload();
+        $this->tmpFolder->delete($this->workspace_path);
+        $this->tmpFolder = null;
+        parent::tearDown();
     }
 
     /**
@@ -124,5 +137,43 @@ class TypeDossierTransformationTest extends PastellTestCase
         $this->assertLastMessage("Transformation terminée");
 
         $this->assertLastDocumentAction('transformation', $info['id_d']);
+    }
+
+    /**
+     * @throws DonneesFormulaireException
+     * @throws NotFoundException
+     * @throws TypeDossierException
+     */
+    public function testEtapeTransformationAfterGlaneur()
+    {
+        $this->createConnectorAndDocument(self::TRANSFORMATION, self::PATH_CONFIG_JSON);
+
+        $glaneurSFTP = $this->getObjectInstancier()->getInstance(GlaneurSFTP::class);
+
+        $glaneurSFTP->setConnecteurInfo(['id_e' => self::ID_E_COL]);
+        $glaneurConfig = $this->getDonneesFormulaireFactory()->getNonPersistingDonneesFormulaire();
+        $glaneurConfig->setTabData([
+            GlaneurConnecteur::TRAITEMENT_ACTIF => '1',
+            GlaneurConnecteur::TYPE_DEPOT => GlaneurConnecteur::TYPE_DEPOT_VRAC,
+            GlaneurConnecteur::FILE_PREG_MATCH => 'fichier: /^(.*).xml$/',
+            GlaneurConnecteur::METADATA_STATIC =>
+                "titre: %fichier%\n
+                iparapheur_type: PADES\n
+                iparapheur_sous_type: Document",
+            GlaneurConnecteur::FLUX_NAME => self::TRANSFORMATION,
+            GlaneurConnecteur::ACTION_OK => 'importation',
+        ]);
+        $glaneurConfig->addFileFromCopy(
+            GlaneurConnecteur::FICHER_EXEMPLE,
+            'pes.zip',
+            __DIR__ . '/../../connecteur/glaneur-sftp/fixtures/HELIOS_SIMU_ALR2_1547544424_844200543.zip'
+        );
+
+        $glaneurSFTP->setConnecteurConfig($glaneurConfig);
+        $id_d = $glaneurSFTP->glanerFicExemple();
+        $this->assertSame("Création du document $id_d", $glaneurSFTP->getLastMessage()[0]);
+
+        $this->triggerActionOnDocument($id_d, "transformation");
+        $this->assertLastMessage("Transformation terminée");
     }
 }